Add files via upload
Question 1 
You are working in a specific company as a data analyst. This company has an available position for country manager and new person applied for this position. He/She has been working for 3 years as a country manager in another company and his/her salary is 133,000$. HR sent you the positions’ salary dataset and ask you if he/she deserve the 133,000$? Consider his\her level on 5.5.